requirements.txt
# requirements.txt
bilibili-api-python>=5.3.0
pydantic>=2.0.0
numpy>=1.21.0
python-dotenv>=1.0.0
aiohttp>=3.8.0